requirements.txt
PyMuPDF>=1.24.0
fpdf2>=2.8.0
rich>=13.0.0
streamlit>=1.30.0
pandas>=2.0.0
plotly>=5.18.0
